move edit-avatar
